+++ src/main/java/com/czareg/simplerestservice/service/calculator/Calculator.java
@@ -1,5 +1,5 @@
 package com.czareg.simplerestservice.service.calculator;
 
 public interface Calculator {
-    int calculate(int followers, int publicReposCount);
+    float calculate(int followers, int publicReposCount);
 }
